Fixed the issue of bidirectional traps
Using the built in networkx methods was still allowing for the creation of maps which had inescapable traps.  This was because paths existed even if they weren't passable.  Therefore it was neccessary to create temporary graphs that had all of the impassible edge removed.  This resolved the issue in a simpler and more effective fashion.